Move energy backup to Delta 3 classic instead of base
