+++ custom_components/ef_ble/eflib/devices/delta3.py
@@ -14,33 +14,5 @@ class Device(delta3_classic.Device):
     usb_ports = pb_field(pb.flow_info_qcusb1, _delta3_base.flow_is_on)
 
     # Energy strategy operation modes
-    energy_strategy_self_powered = pb_field(
-        pb.energy_strategy_operate_mode,
-        lambda x: x.operate_self_powered_open if x else None,
-    )
-    energy_strategy_scheduled = pb_field(
-        pb.energy_strategy_operate_mode,
-        lambda x: x.operate_scheduled_open if x else None,
-    )
-    energy_strategy_tou = pb_field(
-        pb.energy_strategy_operate_mode,
-        lambda x: x.operate_tou_mode_open if x else None,
-    )
-
     async def enable_usb_ports(self, enabled: bool):
         await self._send_config_packet(pd335_sys_pb2.ConfigWrite(cfg_usb_open=enabled))
-
-    async def enable_energy_strategy_self_powered(self, enabled: bool):
-        config = pd335_sys_pb2.ConfigWrite()
-        config.cfg_energy_strategy_operate_mode.operate_self_powered_open = enabled
-        await self._send_config_packet(config)
-
-    async def enable_energy_strategy_scheduled(self, enabled: bool):
-        config = pd335_sys_pb2.ConfigWrite()
-        config.cfg_energy_strategy_operate_mode.operate_scheduled_open = enabled
-        await self._send_config_packet(config)
-
-    async def enable_energy_strategy_tou(self, enabled: bool):
-        config = pd335_sys_pb2.ConfigWrite()
-        config.cfg_energy_strategy_operate_mode.operate_tou_mode_open = enabled
-        await self._send_config_packet(config)
